requirements.txt
Flask
psycopg[binary]
flask-bcrypt
Flask-Login
Flask-Mail
Flask-SQLAlchemy
python-dotenv
